requirements.txt
emoji==2.14.1
pillow==11.1.0
pyperclip==1.9.0
ttkbootstrap==1.10.1
